# server/balance/gin_server.go
package balance

import (
	"encoding/json"
	"fmt"
	"io/ioutil"
	"log"
	"math/rand"
	"net/http"
	"os"
	"regexp"
	"strconv"
	"time"

	_ "transportBalance/docs"
	// docs folder path generated by swag init

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

type Сurrency struct {
	Query struct {
		Apikey       string `json:"apikey"`
		BaseCurrency string `json:"base_currency"`
		Timestamp    int    `json:"timestamp"`
	} `json:"query"`
	Data struct {
		Rub float64 `json:"RUB"`
	} `json:"data"`
}

type ServerGin struct {
	userStorage         IUserStorage
	transactionsStorage ItransactionsStorage
	router              *gin.Engine
	host                string
	port                int
}

type RequestMoveMoney struct {
	UserId int     `json:"userid"`
	Money  float64 `json:"money"`
}

type User struct {
	UserId int `json:"userid"`
}

type RequestUser struct {
	Username string `json:"username"`
	Password string `json:"paswword"`
}

type TransferMoney struct {
	UserIdTo   uint    `json:"UserIdTo"`
	UserIdFrom uint    `json:"UserIdFrom"`
	Money      float64 `json:"Money"`
}

type RequestError struct {
	Error string `json:"error"`
}
type ReserveMoney struct {
	UserId    uint    `json:"userid"`
	Money     float64 `json:"money"`
	ServiceId uint    `json:"serviceId"`
	OrderId   uint    `json:"orderId"`
}

type GenDoc struct {
	Date string `json:"date"`
}

type IChatServer interface {
	Use(userStorage IUserStorage, transactionsStorage ItransactionsStorage)
	Run()
}

func NewServerGin(host string, port int) *ServerGin {
	ServerGin := new(ServerGin)
	ServerGin.host = host
	ServerGin.port = port

	return ServerGin
}

func (serverGin *ServerGin) Use(userStorage IUserStorage, transactionsStorage ItransactionsStorage) {

	serverGin.router = gin.Default()

	serverGin.transactionsStorage = transactionsStorage
	serverGin.userStorage = userStorage
	// Конфигурируем все эндпоинты
	serverGin.router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	serverGin.router.GET("/api/health", heathHandler())
	serverGin.router.POST("/api/user", userHandler(serverGin.userStorage))
	serverGin.router.POST("/api/money", getMoneyUserHadler(serverGin.userStorage))
	serverGin.router.POST("/api/add", addMoneyHandler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/reduce", reduceMoneyHandler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/transfer", transferMoneyHandler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/getmovemoney", getLastTransactionHadler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/reserve", addMoneyToReserveHandler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/reduceReserve", reduceReserveHandler(serverGin.transactionsStorage))
	serverGin.router.POST("/api/docs", createGenDocHandler(serverGin.transactionsStorage))

	// serverGin.router.GET("https://freecurrencyapi.net/api/v2/latest?apikey=d53f1180-94a8-11ec-992b-13a8f6f1bdf9&base_currency=USD",exchangeHandler())

}

func (serverGin *ServerGin) Run() error {
	if serverGin.router == nil {
		return fmt.Errorf("gin не сконфигурирован %v", serverGin.router)
	}

	return serverGin.router.Run(":8080")
}

// heathHandler godoc
// @Summary Show the status of server.
// @Description get the status of server.
// @Tags Get Info
// @Accept json
// @Produce json
// @Router /api/health [get]
func heathHandler() gin.HandlerFunc {
	return func(ctx *gin.Context) {
		ctx.IndentedJSON(http.StatusOK, gin.H{
			"success": true,
			"time":    time.Now().Format(time.RFC3339),
		})

	}
}

// userHandler godoc
// @tags Auth
// @Summary userHandler
// @Description Регистрация пользователя
// @Produce json
// @Param user body RequestUser true "RequestUser"
// @Router /api/user [post]
func userHandler(userStorage IUserStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(RequestUser)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		if !validatenUserName(requestUser.Username) {
			ctx.IndentedJSON(http.StatusForbidden, gin.H{"error": "Invalid Username"})
			return
		}

		if !validatePassword(requestUser.Password) {
			ctx.IndentedJSON(http.StatusForbidden, gin.H{"error": "Invalid Password"})
			return
		}
		userMode, err := userStorage.GetByName(requestUser.Username)
		if err == nil {
			ctx.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "Пользователь существует в базе"})
			return
		}
		userMode, err = userStorage.Create(requestUser.Username, requestUser.Password)
		if err != nil {
			ctx.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
			return
		}
		ctx.IndentedJSON(http.StatusCreated, gin.H{"username": userMode.Username, "money": userMode.Money})
	}

}

// addMoneyHandler godoc
// @tags Balance
// @Summary addMoneyHandler
// @Description Зачислить средства
// @Produce json
// @Param user body RequestMoveMoney true "RequestUser"
// @Router /api/add [post]
func addMoneyHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(RequestMoveMoney)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		_, err := transactionsStorage.AddMoney(uint(requestUser.UserId), requestUser.Money)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": requestUser.UserId, "money": requestUser.Money})
	}
}

// addMoneyToReserveHandler godoc
// @tags Balance
// @Summary addMoneyToReserveHandler
// @Description Резерв денежных средств по определенной услуге
// @Produce json
// @Param user body ReserveMoney true "RequestUser"
// @Router /api/reserve [post]
func addMoneyToReserveHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestReserveMoney := new(ReserveMoney)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestReserveMoney)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		err := transactionsStorage.ReserveMoney(uint(requestReserveMoney.UserId), uint(requestReserveMoney.ServiceId), uint(requestReserveMoney.OrderId), requestReserveMoney.Money)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": requestReserveMoney.UserId, "money": requestReserveMoney.Money, "orderId": requestReserveMoney.OrderId})
	}
}

// reduceReserveHandler godoc
// @tags Balance
// @Summary reduceReserveHandler
// @Description списание средств из Резерва
// @Produce json
// @Param user body ReserveMoney true "RequestUser"
// @Router /api/reduceReserve [post]
func reduceReserveHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestReserveMoney := new(ReserveMoney)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestReserveMoney)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		err := transactionsStorage.ReduceReserveMoney(uint(requestReserveMoney.UserId), uint(requestReserveMoney.ServiceId), uint(requestReserveMoney.OrderId), requestReserveMoney.Money)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": requestReserveMoney.UserId, "money": requestReserveMoney.Money, "orderId": requestReserveMoney.OrderId, "serviceID": requestReserveMoney.ServiceId})
	}
}

// reduceMoneyHandler godoc
// @tags Balance
// @Summary reduceMoneyHandler
// @Description Списать денежные средства
// @Produce json
// @Param user body RequestMoveMoney true "User Data"
// @Router /api/reduce [post]
func reduceMoneyHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(RequestMoveMoney)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		_, err := transactionsStorage.WriteOffMoney(uint(requestUser.UserId), requestUser.Money)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": requestUser.UserId, "money": requestUser.Money})
	}
}

// transferMoneyHandler godoc
// @tags Balance
// @Summary transferMoneyHandler
// @Description Перевести деньги от пользователя к пользователю
// @Produce json
// @Param TransactionsModel body TransferMoney true "User Data"
// @Router /api/transfer [post]
func transferMoneyHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(TransferMoney)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		_, err := transactionsStorage.TransferMoney(requestUser.UserIdFrom, requestUser.UserIdTo, requestUser.Money)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": requestUser.UserIdFrom, "money": requestUser.Money})
	}
}

// createGenDocHandler godoc
// @tags Reports
// @Summary createGenDocHandler
// @Description Получить отчет для Бухгалтерии
// @Produce json
// @Param Date body GenDoc true "RequestUser"
// @Router /api/docs [post]
func createGenDocHandler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(GenDoc)
		var fileDownloadName string
		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		reserveModel, err := transactionsStorage.ListRepots(requestUser.Date)
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		fileDownloadName = createReportCSV(reserveModel)

		ctx.IndentedJSON(http.StatusOK, gin.H{"URLCSV": fmt.Sprintf("http://localhost:8080/download/%s.csv", fileDownloadName)})
	}
}

// getMoneyUserHadler godoc
// @tags Balance
// @Summary getMoneyUserHadler
// @Description Получить данные о балансе
// @Produce json
// @Param user body User true "User Data"
// @Router /api/money [post]
func getMoneyUserHadler(userStorage IUserStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(User)

		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}

		userModel, err := userStorage.GetById(uint(requestUser.UserId))
		if err != nil {
			ctx.IndentedJSON(http.StatusForbidden, err.Error())
			return
		}
		currency := ctx.DefaultQuery("currency", "RUB")
		if currency == "USD" {
			rate := exchangeHandler()
			userModel.Money = userModel.Money / rate
			ctx.IndentedJSON(http.StatusOK, gin.H{"username": userModel.Username, "money": userModel.Money})
			return
		}
		ctx.IndentedJSON(http.StatusOK, gin.H{"userId": userModel.ID, "money": userModel.Money})
	}
}

func getLastTransactionHadler(transactionsStorage ItransactionsStorage) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		requestUser := new(User)
		statusCode, ctx2, checkBadRequest := validateClientRequest(ctx, requestUser)

		if !checkBadRequest {
			ctx.IndentedJSON(statusCode, ctx2)
			return
		}
		page := ctx.DefaultQuery("page", "1")
		filtermoney := ctx.DefaultQuery("filtermoney", "desc")
		filtertime := ctx.DefaultQuery("filtertime", "desc")
		pageint, _ := strconv.Atoi(page)
		transactions, err := transactionsStorage.ListRecords(pageint, filtermoney, filtertime, requestUser.UserId)
		if err != nil {
			ctx.IndentedJSON(http.StatusBadRequest, err.Error())
			return
		}
		ctx.IndentedJSON(http.StatusOK, transactions)
	}
}

func exchangeHandler() float64 {
	resp, err := http.Get("https://freecurrencyapi.net/api/v2/latest?apikey=d53f1180-94a8-11ec-992b-13a8f6f1bdf9&base_currency=USD")
	if err != nil {
		log.Fatalln(err)
	}
	dd := new(Сurrency)
	body, err := ioutil.ReadAll(resp.Body)
	if err != nil {
		fmt.Println(err.Error())
	}
	json.Unmarshal(body, dd)
	return dd.Data.Rub
}

func validatenUserName(userName string) bool {
	var validPath = regexp.MustCompile(`^[a-zA-Z0-9]+$`).MatchString
	return validPath(userName) && len(userName) < 16
}

func validatePassword(password string) bool {
	return len(password) < 32
}

func validateClientRequest(ctx *gin.Context, requestData interface{}) (int, interface{}, bool) {

	err := ctx.BindJSON(&requestData)
	if err != nil {

		return http.StatusBadRequest, gin.H{"error": "Не содержит поля в запросе"}, false
	}
	return http.StatusOK, gin.H{"error": ""}, true
}

func createReportCSV(reserveModel []ReserveModel) string {

	generator := rand.New(rand.NewSource(time.Now().UnixNano()))
	n := strconv.Itoa(generator.Int())
	file, err := os.Create("./download/" + n + ".csv")

	if err != nil {
		fmt.Println("Unable to create file:", err)
		return ""
	}
	defer file.Close()
	for _, value := range reserveModel {
		file.WriteString(fmt.Sprintf("%s;%s\n", value.ServiceId, value.Money))
	}

	return file.Name()
}